requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Telegram Bot
python-telegram-bot==21.7

# LLM Providers
openai==1.54.0
anthropic==0.39.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# Scheduler
apscheduler==3.10.4

# Google Calendar (Optional)
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# NLP & Date Parsing
dateparser==1.2.0
python-dateutil==2.9.0

# Logging
structlog==24.4.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Testing (Optional)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
